requirements.txt
tox~=4.6.0
black~=23.3.0
mypy~=1.3.0
pyinstaller~=5.12.0
types-setuptools~=68.0.0.0

pynng~=0.7.2
pyserial~=3.5
pikoder_ppm_encoder~=0.0.2